Restarted a few times, really ugly but it works!
I will have to go back and clean up the code. It's not pretty
but it works! Some additional functionality would be nice; I
still need to add some non-ascii (<128 anyway) characters, and
perhaps distinguish between a file and an empty directory, things
like that. Still, happy that it builds (without warnings!) and
prints out correctly.